Moved GeoOptions in to specific package.
