+++ java/client/src/main/java/glide/api/commands/GeospatialIndicesBaseCommands.java
@@ -1,8 +1,8 @@
 /** Copyright GLIDE-for-Redis Project Contributors - SPDX Identifier: Apache-2.0 */
 package glide.api.commands;
 
-import glide.api.models.commands.GeoAddOptions;
-import glide.api.models.commands.GeospatialData;
+import glide.api.models.commands.geospatial.GeoAddOptions;
+import glide.api.models.commands.geospatial.GeospatialData;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
