+++ java/client/src/main/java/glide/api/models/commands/geospatial/GeospatialData.java
@@ -1,5 +1,5 @@
 /** Copyright GLIDE-for-Redis Project Contributors - SPDX Identifier: Apache-2.0 */
-package glide.api.models.commands;
+package glide.api.models.commands.geospatial;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
