+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "@aws/glide-for-redis": "^0.1.2"
+  }
+}
